Fixing event persistence

+++ cd2024bfs5g1-model/src/main/resources/db/migration/V19__alter_event_table.sql
@@ -1,1 +1,1 @@
-ALTER TABLE event ADD price INTEGER;
+ALTER TABLE event ADD price NUMERIC(10,2);
